Rename TestApplication.Wcf.Client.Core to TestApplication.Wcf.Client.DotNet
