+++ OpenTelemetry.AutoInstrumentation.sln
@@ -117,13 +117,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTelemetry.AutoInstrumen
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApplication.StrongNamed", "test\test-applications\integrations\TestApplication.StrongNamed\TestApplication.StrongNamed.csproj", "{505A5B7B-930E-42D3-909F-E97C7C247FCF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApplication.Wcf.Client.Core", "test\test-applications\integrations\TestApplication.Wcf.Client.Core\TestApplication.Wcf.Client.Core.csproj", "{5BFB9A97-A83F-4719-A35B-6421364B9820}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApplication.Wcf.Client.DotNet", "test\test-applications\integrations\TestApplication.Wcf.Client.DotNet\TestApplication.Wcf.Client.DotNet.csproj", "{5BFB9A97-A83F-4719-A35B-6421364B9820}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApplication.Wcf.Client.NetFramework", "test\test-applications\integrations\TestApplication.Wcf.Client.NetFramework\TestApplication.Wcf.Client.NetFramework.csproj", "{ACB51ACB-82E6-4CA8-B233-2C05A5CEDE01}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApplication.Wcf.Server.NetFramework", "test\test-applications\integrations\TestApplication.Wcf.Server.NetFramework\TestApplication.Wcf.Server.NetFramework.csproj", "{F7ACC09E-899F-4D47-BA1D-4009FB76F6AA}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApplication.Wcf.Shared", "test\test-applications\integrations\TestApplication.Wcf.Shared\TestApplication.Wcf.Shared.csproj", "{86B61808-6B61-465E-8451-84DC31935C03}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApplication.Logs", "test\test-applications\integrations\TestApplication.Logs\TestApplication.Logs.csproj", "{8C52B9DD-880B-4CD5-AB18-A4C462C714E9}"
 EndProject
@@ -535,18 +533,6 @@ Global
 		{F7ACC09E-899F-4D47-BA1D-4009FB76F6AA}.Release|x64.Build.0 = Release|x64
 		{F7ACC09E-899F-4D47-BA1D-4009FB76F6AA}.Release|x86.ActiveCfg = Release|x86
 		{F7ACC09E-899F-4D47-BA1D-4009FB76F6AA}.Release|x86.Build.0 = Release|x86
-		{86B61808-6B61-465E-8451-84DC31935C03}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Debug|Any CPU.Build.0 = Debug|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Debug|x64.ActiveCfg = Debug|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Debug|x64.Build.0 = Debug|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Debug|x86.ActiveCfg = Debug|x86
-		{86B61808-6B61-465E-8451-84DC31935C03}.Debug|x86.Build.0 = Debug|x86
-		{86B61808-6B61-465E-8451-84DC31935C03}.Release|Any CPU.ActiveCfg = Release|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Release|Any CPU.Build.0 = Release|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Release|x64.ActiveCfg = Release|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Release|x64.Build.0 = Release|x64
-		{86B61808-6B61-465E-8451-84DC31935C03}.Release|x86.ActiveCfg = Release|x86
-		{86B61808-6B61-465E-8451-84DC31935C03}.Release|x86.Build.0 = Release|x86
 		{8C52B9DD-880B-4CD5-AB18-A4C462C714E9}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{8C52B9DD-880B-4CD5-AB18-A4C462C714E9}.Debug|Any CPU.Build.0 = Debug|x64
 		{8C52B9DD-880B-4CD5-AB18-A4C462C714E9}.Debug|x64.ActiveCfg = Debug|x64
@@ -708,7 +694,6 @@ Global
 		{5BFB9A97-A83F-4719-A35B-6421364B9820} = {E409ADD3-9574-465C-AB09-4324D205CC7C}
 		{ACB51ACB-82E6-4CA8-B233-2C05A5CEDE01} = {E409ADD3-9574-465C-AB09-4324D205CC7C}
 		{F7ACC09E-899F-4D47-BA1D-4009FB76F6AA} = {E409ADD3-9574-465C-AB09-4324D205CC7C}
-		{86B61808-6B61-465E-8451-84DC31935C03} = {E409ADD3-9574-465C-AB09-4324D205CC7C}
 		{8C52B9DD-880B-4CD5-AB18-A4C462C714E9} = {E409ADD3-9574-465C-AB09-4324D205CC7C}
 		{1DC12784-CD32-400D-AE13-A637BFA66B1B} = {82A3CE96-0935-45E5-A9AA-A93A5B63500B}
 		{28879377-2043-4370-A8DB-B04464B6E552} = {E409ADD3-9574-465C-AB09-4324D205CC7C}
