+++ test/test-applications/integrations/TestApplication.Wcf.Client.DotNet/IStatusServiceContract.cs
@@ -1,4 +1,4 @@
-// <copyright file="WcfCoreTests.cs" company="OpenTelemetry Authors">
+// <copyright file="IStatusServiceContract.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,18 +14,14 @@
 // limitations under the License.
 // </copyright>
 
-// This test won't work outside of windows as it need the server side which is .NET Framework only.
-#if NET6_0_OR_GREATER && _WINDOWS
-using Xunit.Abstractions;
+using System.ServiceModel;
+using System.Threading.Tasks;
 
-namespace IntegrationTests;
+namespace TestApplication.Wcf.Client.DotNet;
 
-public class WcfCoreTests : WcfTestsBase
+[ServiceContract(Namespace = "http://opentelemetry.io/", Name = "StatusService", SessionMode = SessionMode.Allowed)]
+public interface IStatusServiceContract
 {
-    public WcfCoreTests(ITestOutputHelper output)
-        : base("Wcf.Client.Core", output)
-    {
-    }
+    [OperationContract]
+    Task<StatusResponse> PingAsync(StatusRequest request);
 }
-
-#endif
