+++ test/test-applications/integrations/TestApplication.Wcf.Client.DotNet/StatusRequest.cs
@@ -1,4 +1,4 @@
-// <copyright file="StatusResponse.cs" company="OpenTelemetry Authors">
+// <copyright file="StatusRequest.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,14 +14,13 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using System.Runtime.Serialization;
 
-namespace TestApplication.Wcf.Client.Core;
+namespace TestApplication.Wcf.Client.DotNet;
 
 [DataContract]
-public class StatusResponse
+public class StatusRequest
 {
     [DataMember]
-    public DateTimeOffset ServerTime { get; set; }
+    public string? Status { get; set; }
 }
