+++ test/test-applications/integrations/TestApplication.Wcf.Client.DotNet/StatusServiceClient.cs
@@ -18,7 +18,7 @@ using System.ServiceModel;
 using System.ServiceModel.Channels;
 using System.Threading.Tasks;
 
-namespace TestApplication.Wcf.Client.Core;
+namespace TestApplication.Wcf.Client.DotNet;
 
 public class StatusServiceClient : ClientBase<IStatusServiceContract>, IStatusServiceContract
 {
