+++ test/test-applications/integrations/TestApplication.Wcf.Client.DotNet/TestApplication.Wcf.Client.DotNet.csproj
